+++ src/aihwkit/simulator/rpu_base_src/rpu_base_devices.cpp
@@ -349,7 +349,9 @@ void declare_rpu_devices(py::module &m) {
       .def_readwrite("out_scale", &RPU::IOMetaParameter<T>::out_scale)
       .def_readwrite("out_sto_round", &RPU::IOMetaParameter<T>::out_sto_round)
       .def_readwrite("w_noise", &RPU::IOMetaParameter<T>::w_noise)
-      .def_readwrite("w_noise_type", &RPU::IOMetaParameter<T>::w_noise_type);
+      .def_readwrite("w_noise_type", &RPU::IOMetaParameter<T>::w_noise_type)
+      .def_readwrite("ir_drop", &RPU::IOMetaParameter<T>::ir_drop)
+      .def_readwrite("ir_drop_g_ratio", &RPU::IOMetaParameter<T>::ir_drop_Gw_div_gmax);
 
   py::class_<RPU::DriftParameter<T>>(m, "DriftParameter")
       .def(py::init<>())
@@ -567,10 +569,12 @@ void declare_rpu_devices(py::module &m) {
       .def_readwrite(
           "transfer_every_vec", &TransferParam::transfer_every_vec) // can this be filled?
       .def_readwrite("units_in_mbatch", &TransferParam::units_in_mbatch)
-      .def_readwrite("n_cols_per_transfer", &TransferParam::n_cols_per_transfer)
+      .def_readwrite("n_reads_per_transfer", &TransferParam::n_reads_per_transfer)
       .def_readwrite("with_reset_prob", &TransferParam::with_reset_prob)
-      .def_readwrite("random_column", &TransferParam::random_column)
+      .def_readwrite("random_selection", &TransferParam::random_selection)
+      .def_readwrite("transfer_columns", &TransferParam::transfer_columns)
       .def_readwrite("transfer_lr", &TransferParam::transfer_lr)
+      .def_readwrite("fast_lr", &TransferParam::fast_lr)
       .def_readwrite("transfer_lr_vec", &TransferParam::transfer_lr_vec)
       .def_readwrite("scale_transfer_lr", &TransferParam::scale_transfer_lr)
       .def_readwrite("transfer_forward", &TransferParam::transfer_io)