+++ src/rpucuda/cuda/rpucuda_expstep_test.cpp
@@ -38,8 +38,7 @@ template <typename Tio> class RPUCudaExpstepStateTestFixture : public ::testing:
 public:
   void SetUp() {
 
-    context_container = RPU::make_unique<CudaContext>(-1, false);
-    context = &*context_container;
+    context = &context_container;
 
     x_size = 100;
     d_size = 110;
@@ -140,7 +139,7 @@ public:
     d_vec2_batch.resize(m_batch * d_size);
   }
 
-  std::unique_ptr<CudaContext> context_container;
+  CudaContext context_container{-1, false};
   CudaContext *context;
 
   std::unique_ptr<RPUPulsed<num_t>> layer_pulsed;