+++ src/rpucuda/rpu_vector_device.h
@@ -199,6 +199,7 @@ public:
   void doSparseUpdate(
       T **weights, int i, const int *x_signed_indices, int x_count, int d_sign, RNG<T> *rng)
       override;
+  void doDenseUpdate(T **weights, int *coincidences, RNG<T> *rng) override;
 
 protected:
   void populate(const VectorRPUDeviceMetaParameter<T> &par, RealWorldRNG<T> *rng);